configurando compilador ts
