+++ .gitignore
@@ -9,6 +9,7 @@ npm-debug.*
 *.mobileprovision
 *.orig.*
 web-build/
+.idea/
 
 # macOS
 .DS_Store
