+++ App.tsx
@@ -1,7 +1,5 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
-import LoginScreen from "./src/screens/login";
-import MainNavigation from './src/navigations';
+import { StyleSheet } from 'react-native';
+import MainNavigation from './src/navigations/StackContainer';
 import 'react-native-gesture-handler';
 
 export default function App() {
