+++ package.json
@@ -4,9 +4,11 @@
   "main": "node_modules/expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
+    "debug": "npx expo start --tunnel",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web"
+    "web": "expo start --web",
+    "check": "tsc --noEmit"
   },
   "dependencies": {
     "@expo-google-fonts/ubuntu": "^0.2.3",
