+++ src/screens/common/ErrorScreen.tsx
@@ -1,0 +1,3 @@
+export default function ErrorScreen(){
+    
+}
