+++ src/screens/common/LoadingScreen.tsx
@@ -1,0 +1,3 @@
+export default function LoadingScreen(){
+
+}
