+++ src/screens/home/RegisterScreen.tsx
@@ -1,0 +1,3 @@
+export default function RegisterScreen(){
+    
+}
