+++ tsconfig.json
@@ -1,7 +1,14 @@
 {
   "extends": "expo/tsconfig.base",
+  "include": ["**/*.tsx", "src/react-app-env.d.ts"],
   "compilerOptions": {
-    "jsx": "react",
-    "strict": true
+    "jsx": "react-jsx",
+    "strict": true,
+    "baseUrl": "src",
+    "paths": {
+      "@assets/*": ["../assets/*"],
+      "@common/*": ["screens/common/*"],
+      "@screens/*": ["screens/*"]
+    }
   }
 }
